Add into_ref(self) -> &'a T for InternArena<'a, T>
This allows you to store arena-interned values into structures inline
(i.e. without `&'a InternArena<'a, T>`, which is just wasteful), and
still be able to borrow the values inside of the interned value without
providing a reference itself. For example:
```rs
struct Bar {
    baz: String,
}

struct Foo<'a>(InternArena<'a, Bar>);

impl<'a> Foo<'a> {
    pub fn get_baz(self) -> &'a str {
        self.0.into_ref().baz
    }
}
```

Without this addition, doing this requires a `&self` in order for the
borrow checker to be satisfied, which to be honest is wasteful, when
this type can be `Copy` anyway. This is especially useful when using
wrapper types to enforce API boundaries.